[Bronze I] Title: 피보나치 수 2, Time: 164 ms, Memory: 17696 KB -BaekjoonHub

+++ 백준/Bronze/2748. 피보나치 수 2/피보나치 수 2.java
@@ -3,14 +3,14 @@ import java.util.*;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        long n = Integer.parseInt(sc.nextLine());
+        int n = Integer.parseInt(sc.nextLine());
 
-        long[] fibo = new long[(int) (n + 1)];
+        long[] fibo = new long[n + 1];
         fibo[0] = 0; fibo[1] = 1;
 
         for (int i = 2; i < fibo.length; i++) {
             fibo[i] = fibo[i - 1] + fibo[i - 2];
         }
-        System.out.print(fibo[(int) n]);
+        System.out.print(fibo[n]);
     }
 }
